kinda works but its slow

+++ shell/shell.js
@@ -2,6 +2,7 @@
 const path = require('path');
 const pty = require('node-pty');
 const EventEmitter = require('events');
+const child_process = require('child_process');
 const magicToken = `\x33[JOELMAGIC${Math.random()}]`;
 const magicString = magicToken + '\r\n';
 let magicCallback = null;
@@ -9,43 +10,15 @@ let magicCallback = null;
 class Shell extends EventEmitter {
   constructor() {
     super();
+    this.rows = 80;
+    this.cols = 24;
+    this.env = {...process.env};
+    this.cwd = process.cwd();
     this.inCommand = false;
     this.ignoreText = '';
     this.evaluating = false;
     this.evaluateResult = '';
     this.commandQueue = Promise.resolve();
-    this.shell = pty.spawn('bash', [path.join(__dirname, 'proxy.sh'), magicToken], {
-      rows: 80,
-      cols: 24,
-      cwd: process.cwd(),
-      name: 'xterm-256color',
-      handleFlowControl: true,
-      encoding: null,
-    });
-    this.shell.onData(data => {
-      const dataString = data.toString();
-      let index = 0;
-      for (index = 0; index < this.ignoreText.length && index < dataString.length; index++) {
-        if (dataString[index] !== this.ignoreText[index]) {
-          this.ignoreText = '';
-          break;
-        }
-      }
-      if (index > 0) {
-        data = data.slice(index);
-        this.ignoreText = this.ignoreText.slice(index);
-      }
-      if (magicCallback && data.slice(data.length - magicString.length).toString() === magicString) {
-        data = data.slice(0, -magicString.length);
-        magicCallback();
-        magicCallback = null;
-      }
-      if (this.evaluating)
-        this.evaluateResult += data;
-      else
-        this.emit('data', data);
-    });
-    this.shell.onExit(event => this.emit('close', event.exitCode, event.signal));
   }
 
   /**
@@ -55,13 +28,49 @@ class Shell extends EventEmitter {
     return this.commandQueue = this.commandQueue.then(() => this._innerRunCommand(command));
   }
 
+  resize(cols, rows) {
+    this.rows = rows;
+    this.cols = cols;
+    if (this.shell)
+      this.shell.resize(cols, rows);
+  }
+
   /**
    * @param {string} command
    */
   async _innerRunCommand(command) {
-    this.ignoreText = command + '\r\n';
-    this.shell.write(command + '\n');
-    await new Promise(x => magicCallback = x);
+    this.shell = pty.spawn('node', [path.join(__dirname, '..', 'shjs', 'wrapper.js'), command, magicToken], {
+      env: this.env,
+      rows: this.rows,
+      cols: this.cols,
+      cwd: this.cwd,
+      name: 'xterm-256color',
+      handleFlowControl: true,
+      encoding: null,
+    });
+    let extraData = '';
+    let inExtraData = false;
+    this.shell.onData(data => {
+      if (!inExtraData && data.slice(data.length - magicString.length).toString() === magicString) {
+        data = data.slice(0, -magicString.length);
+        inExtraData = true;
+      }
+      if (inExtraData)
+        extraData += data;
+      else
+        this.emit('data', data);
+    });
+    await new Promise(x => this.shell.onExit(x));
+    this.shell = null;
+    if (extraData.length) {
+      const changes = JSON.parse(extraData);
+      if (changes.cwd)
+        this.cwd = changes.cwd;
+      if (changes.env) {
+        for (const key in changes.env)
+          this.env[key] = changes.env[key];
+      }
+    }
   }
 
   /**
@@ -69,24 +78,28 @@ class Shell extends EventEmitter {
    * @return {Promise<string>}
    */
   async evaluate(code) {
-    let result;
-    this.commandQueue = this.commandQueue.then(async () => {
-      this.evaluating = true;
-      await this._innerRunCommand(code);
-      this.evaluating = false;
-      result = this.evaluateResult;
-      this.evaluateResult = '';
+    // return '';
+    const child = child_process.spawn('node', [path.join(__dirname, '..', 'shjs', 'wrapper.js'), code], {
+      env: this.env,
+      cwd: this.cwd,
+      stdio: ['pipe', 'pipe', 'pipe'],
     });
-    await this.commandQueue;
-    return result;
+    const data = [];
+    child.stdin.end();
+    child.stderr.on('data', d => data.push(d));
+    child.stdout.on('data', d => data.push(d));
+    await new Promise(x => child.on('exit', x));
+    return data.join('');
   }
 
   sendRawInput(input) {
-    this.shell.write(input);
+    if (this.shell)
+      this.shell.write(input);
   }
 
   close() {
-    this.shell.kill();
+    if (this.shell)
+      this.shell.kill();
   }
 }
